pyproject.toml
[tool.poetry]
name = "shopify remix micro service"
version = "0.1.0"
description = "This is a Micro Service Project for Shopify Remix App. It uses Django and postgres."
authors = ["Bikin Shrestha <bikin@qodemedia.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.4"
djangorestframework = "^3.15.1"
pillow = "^10.3.0"
drf-yasg = "^1.21.7"
python-dotenv = "^1.0.1"
whitenoise = "^6.6.0"
psycopg2-binary = "^2.9.9"
ShopifyAPI = "^12.5.0"
django-cors-headers = "^4.3.1"

[tool.poetry.group.dev.dependencies]
djlint = "^1.34.1"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
